+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Task to reset the stripe payment methods in test mode
 - Validate on server side the reservation of slots restricted to subscribers
 − Unified and documented upgrade exit codes
+- During setup, ask for the name of the external network and create it, if it does not already exists
 - Fix a bug: build status badge is not working
 - Fix a bug: unable to set date formats during installation
 - Fix a bug: unable to cancel the upgrade before it begins